+++ packages/web/app/components/SearchBar/AllPatientsSearchBar.jsx
@@ -1,6 +1,5 @@
 import React, { useState } from 'react';
 import Box from '@material-ui/core/Box';
-import { startCase } from 'lodash';
 import { useDateTime } from '@tamanu/ui-components';
 import styled from 'styled-components';
 import { CustomisableSearchBarWithPermissionCheck } from './CustomisableSearchBar';
@@ -19,6 +18,7 @@ import { SearchBarCheckField } from './SearchBarCheckField';
 import { TranslatedText } from '../Translation/TranslatedText';
 import { SEX_LABELS, SEX_VALUES } from '@tamanu/constants';
 import { useSettings } from '../../contexts/Settings';
+import { AdditionalSearchField } from './AdditionalSearchField';
 
 const TwoColumnsField = styled(Box)`
   grid-column: span 2;
@@ -34,66 +34,6 @@ const SexLocalisedField = styled(LocalisedField)`
 const VillageLocalisedField = styled(LocalisedField)`
   font-size: 11px;
 `;
-
-const REFERENCE_DATA_FIELDS = new Set([
-  'nationalityId',
-  'countryId',
-  'divisionId',
-  'subdivisionId',
-  'medicalAreaId',
-  'nursingZoneId',
-  'settlementId',
-  'ethnicityId',
-  'occupationId',
-  'religionId',
-  'patientBillingTypeId',
-  'countryOfBirthId',
-  'insurerId',
-  'secondaryVillageId',
-]);
-
-const getReferenceDataType = fieldName => fieldName.replace(/Id$/, '');
-
-const getFieldLabel = fieldName => startCase(fieldName.replace(/Id$/, ''));
-
-const ReferenceDataSearchField = ({ fieldName }) => {
-  const referenceType = getReferenceDataType(fieldName);
-  const suggester = useSuggester(referenceType);
-  return (
-    <LocalisedField
-      component={AutocompleteField}
-      name={fieldName}
-      label={
-        <TranslatedText
-          stringId={`general.localisedField.${fieldName}.label.short`}
-          fallback={getFieldLabel(fieldName)}
-        />
-      }
-      suggester={suggester}
-      size="small"
-    />
-  );
-};
-
-const TextSearchField = ({ fieldName }) => (
-  <LocalisedField
-    component={SearchField}
-    name={fieldName}
-    label={
-      <TranslatedText
-        stringId={`general.localisedField.${fieldName}.label.short`}
-        fallback={getFieldLabel(fieldName)}
-      />
-    }
-  />
-);
-
-const AdditionalSearchField = ({ fieldName }) => {
-  if (REFERENCE_DATA_FIELDS.has(fieldName)) {
-    return <ReferenceDataSearchField fieldName={fieldName} />;
-  }
-  return <TextSearchField fieldName={fieldName} />;
-};
 
 export const AllPatientsSearchBar = React.memo(({ onSearch, searchParameters }) => {
   const { getCurrentDate } = useDateTime();
